// src/server/api/models/accounting/models/PageTemplateCurrencyRate.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CurrencyRate } from './CurrencyRate';
import type { Direction } from './Direction';

export type PageTemplateCurrencyRate = {
    /**
     * The page number requested. A zero-based page index (0..N)
     */
    page?: number;
    /**
     * The size of the page to be returned
     */
    size?: number;
    /**
     * The total number of pages available
     */
    totalPages?: number;
    /**
     * The total number of records available
     */
    totalCount?: number;
    /**
     * The field used to sort the response
     */
    sort: string;
    /**
     * The sort direction
     */
    order: Direction;
    /**
     * The records returned
     */
    content: Array<CurrencyRate>;
};